requirements.txt
apache-airflow==2.7.1
praw
configparser==7.2.0
